+++ dotDoc.LocalIpc/Serializers/DefaultSerializer.cs
@@ -2,33 +2,39 @@
 // This file is licensed to you under the MIT license.
 // See the License.txt file in the solution root for more information.
 
-using Newtonsoft.Json;
+using System;
 using System.Text;
+using System.Text.Json;
 
 namespace DotDoc.LocalIpc.Serializers
 {
     /// <inheritdoc/>
     public class DefaultSerializer : ISerializer
     {
-        private static readonly JsonSerializerSettings settings = new ()
-        {
-            TypeNameHandling = TypeNameHandling.All
-        };
-
         /// <inheritdoc/>
-        public byte[] Serialize(object value)
+        public byte[] Serialize<T>(T value)
         {
-            string json = JsonConvert.SerializeObject(value, settings);
-            byte[] bytes = Encoding.UTF8.GetBytes(json);
-            return bytes;
+            // If the value is not null then use the type of the value for serialization.
+            // For example if we pass an object that contains an int, then use int as the serialization type.
+            Type type = value?.GetType() ?? typeof(T);
+
+            // Create an anonymous type containing the type of the value and the value and serialize this.
+            var valueWrapper = new { TypeName = type.AssemblyQualifiedName, Value = value };
+            string json = JsonSerializer.Serialize(valueWrapper);
+            return Encoding.UTF8.GetBytes(json);
         }
 
         /// <inheritdoc/>
         public T Deserialize<T>(byte[] bytes)
         {
-            string json = Encoding.UTF8.GetString(bytes);
-            T value = JsonConvert.DeserializeObject<T>(json, settings);
-            return value;            
+            using JsonDocument jsonDocument = JsonDocument.Parse(bytes);
+
+            string typeName = jsonDocument.RootElement.GetProperty("TypeName").GetString();
+            Type type = Type.GetType(typeName, true, false);
+
+            // Get the raw text of the value property and deserialize this as the type specified in TypeName.
+            string valueJson = jsonDocument.RootElement.GetProperty("Value").GetRawText();
+            return (T)JsonSerializer.Deserialize(valueJson, type);
         }
     }
 }