+++ dotDoc.LocalIpc/Serializers/ISerializer.cs
@@ -15,14 +15,15 @@ namespace DotDoc.LocalIpc.Serializers
         /// <summary>
         /// Serialize the specified object into a byte array.
         /// </summary>
+        /// <typeparam name="T">The type of object to serialize.</typeparam>
         /// <param name="value">The object to serialize.</param>
         /// <returns>A byte array containing the serialized object.</returns>
-        public byte[] Serialize(object value);
+        public byte[] Serialize<T>(T value);
 
         /// <summary>
         /// Deserialize the specified byte array into an object.
         /// </summary>
-        /// <typeparam name="T">The type of object to deserialize to.</typeparam>
+        /// <typeparam name="T">The type of object to deserialize.</typeparam>
         /// <param name="bytes">The byte array to deserialize.</param>
         /// <returns>A instance of <see cref="{T}"/>.</returns>
         public T Deserialize<T>(byte[] bytes);